move logic into screen
